Remove interface for DecisionTreeFactory
